split up stacked pd into two files
